Sped up create page test
